+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.8.8]
+
+### Update
+
+- Added `DevTools` interface to access the internal state of the game
+- Added `Namespace.has`, `Namespace.keys`, `Namespace.values`, `Namespace.entries`
+- Added `Storable.createNamespace`
+
 ## [0.8.7]
 
 ### Fixed